Migrate Local Find / Widget data out of old reducer
